Split wind.rs (630 lines) into wind/ subdirectory
Split crates/simulation/src/wind.rs into focused modules under wind/:
- types.rs (106 lines): WindState resource, prevailing_direction_for_zone
- systems.rs (149 lines): update_wind system, RNG helpers, constants
- plugin.rs (17 lines): WindPlugin
- tests.rs (375 lines): all unit and integration tests
- mod.rs (9 lines): thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
