+++ crates/simulation/src/wind/plugin.rs
@@ -1,0 +1,17 @@
+use bevy::prelude::*;
+
+use super::systems::update_wind;
+use super::types::WindState;
+
+pub struct WindPlugin;
+
+impl Plugin for WindPlugin {
+    fn build(&self, app: &mut App) {
+        app.init_resource::<WindState>().add_systems(
+            FixedUpdate,
+            update_wind
+                .after(crate::imports_exports::process_trade)
+                .in_set(crate::SimulationSet::Simulation),
+        );
+    }
+}
